tower building should work
towerBase.cs now detects if a button is hit with the raycast and executed the BuildTower function.

buttons themselves don't do anything because adding an action to a button via script is for some reason very hard.

with this approach, buttons are 'clickable' in all camera modes, so that bug is now also gone.